+++ src/utils/clients/voting.ts
@@ -1,9 +1,9 @@
 import { AnyVotingReputationClient, Extension } from '@colony/colony-js';
-import networkClient from '../networkClient';
+import { getCachedColonyClient } from './colony';
 
 export const getVotingClient = async (
   colonyAddress: string,
 ): Promise<AnyVotingReputationClient> => {
-  const colonyClient = await networkClient.getColonyClient(colonyAddress);
+  const colonyClient = await getCachedColonyClient(colonyAddress);
   return await colonyClient.getExtensionClient(Extension.VotingReputation);
 };
